Update fifo.v
Using blocking statement inside always block to sequentially read and write as per the queue's status. If the queue is empty at first and read and write operation both are given then first we will write then read. 
If the queue is full then we will first carry out the read operation then execute the write operation. 
This will manage the edge cases.